queue subsystem: further improve robustness
This now also guards the first write to the new file after
queue rollover. That permits rsyslog to clean out the file
after restart.

Also, a robustness write is now done when a queue file is
fully processed and deleted. Otherwise, the .qi file contains
a non-longer-existing read file number.