queue subsystem: .qi file is automically rewritten
previously, a kill -9 during the .qi write could keep the file
in inconsistent state. Now, we first write a temp file, and
(automically) rename it to "the real thing". So if something happens
during writing the .qi, at least the old state still is consistent.

This is inspired by the way the pid file is handled sind v8.19.