queue subsystem: remove force-persist at begin of shutdown sequence
first of all, this creates race issues and as such is not clean. We
also do no longer need it, as the changes we made to file rollover
cover this use case as well. So instead if working on proper sync,
which is complex, we remove this operation again. Note that it
was NOT present in any released version, just the has been inside
the code the past couple of days while we work on queue robustness.